Add files via upload
A Music Library System is a software application used to manage the storage and retrieval of music
tracks and albums. System need to maintain information such as track title, artist, album, genre, duration,
release year, and any other relevant attributes. System should able to add new music track, remove the
track, update the information of the track, search the specific track also it should display the information
of all the track which are present in the database. Use the Concept of Circular Linked list to implement the
above system.