Flight.cpp
There are flight paths between cities. If there is a flight between city A and city B then there is an 
edge between the cities. The cost of the edge can be the time that flight takes to reach city B from A or 
the amount of fuel used for the journey. The node can be represented by airport name or name of the 
city. Use adjacency list representation of the graph and adjacency matrix representation of the graph. 
Justify the storage representation used.